// src/client/types.gen.ts
export type GetFuncionalidades = {
  /**
   * Chave
   * Chave de acesso autoincrementada
   */
  chave?: number | null
  /**
   * Nome
   * Nome amigável da funcionalidades (ex: 'Excluir Cliente')
   */
  nome?: string | null
  /**
   * Escopo
   * Identificador único no formato 'recurso:ação' (ex: 'client:delete')
   */
  escopo?: string | null
  /**
   * Caminho
   * Endpoint da API (ex: '/api/client/{id}')
   */
  caminho?: string | null
  /**
   * Metodo
   * Método HTTP (GET, POST, PUT, DELETE, etc)
   */
  metodo?: string | null
  /**
   * Descricao
   * Descrição detalhada do propósito da funcionalidades
   */
  descricao?: string | null
  /**
   * Categoria
   * Grupo funcional (ex: 'CRM', 'Financeiro', 'Relatórios')
   */
  categoria?: string | null
  /**
   * Sensivel
   * Indica se é uma operação crítica/cuidado extra necessário
   */
  sensivel?: boolean | null
  /**
   * Versao
   * Versão da funcionalidades para controle de evolução
   */
  versao?: number | null
  /**
   * Id
   * Identificador único universal (UUIDv4) da entidade
   */
  id?: string
  /**
   * Criado Em
   * Data e hora UTC da criação do registro
   */
  criado_em?: string | null
  /**
   * Atualizado Em
   * Data e hora UTC da última atualização do registro (null se nunca foi atualizado)
   */
  atualizado_em?: string | null
}
// ...
export type GetPapeis = {
  /**
   * Nome
   * Nome identificador único do papeis (ex: 'Admin', 'Gerente')
   */
  nome?: string | null
  /**
   * Ativo
   * Indica se o papeis está ativo e disponível para atribuição
   */
  ativo?: boolean | null
  /**
   * Descricao
   * Descrição detalhada das permissões e responsabilidades do papeis
   */
  descricao?: string | null
  /**
   * Id
   * Identificador único universal (UUIDv4) da entidade
   */
  id?: string
  /**
   * Criado Em
   * Data e hora UTC da criação do registro
   */
  criado_em?: string | null
  /**
   * Atualizado Em
   * Data e hora UTC da última atualização do registro (null se nunca foi atualizado)
   */
  atualizado_em?: string | null
}
// ...
export type GetPapeisFuncionalidades = {
  /**
   * Papel Id
   * Chave estrangeira referenciando o ID do Papeis
   */
  papel_id?: string | null
  /**
   * Funcionalidade Id
   */
  funcionalidade_id?: string | null
  /**
   * Id
   * Identificador único universal (UUIDv4) da entidade
   */
  id?: string
  /**
   * Criado Em
   * Data e hora UTC da criação do registro
   */
  criado_em?: string | null
  /**
   * Atualizado Em
   * Data e hora UTC da última atualização do registro (null se nunca foi atualizado)
   */
  atualizado_em?: string | null
  funcionalidades?: GetFuncionalidades | null
  papeis?: GetPapeis | null
}